# docs/architecture/openapi.yaml
openapi: 3.0.3
info:
  title: DataShare API
  description: |
    Secure file transfer platform API for freelancers and small businesses.
    Enables file upload (authenticated or anonymous), temporary download link
    generation, tag/password management and file history.
  version: 1.0.0
  contact:
    name: DataShare Team
servers:
  - url: https://localhost/api
    description: Local Docker Compose (via Nginx reverse proxy)
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token (15min TTL by default, see JWT_EXPIRES_IN)
  parameters:
    FileId:
      in: path
      name: fileId
      required: true
      schema:
        type: string
        format: uuid
  schemas:
...
    Error:
      type: object
      required: [statusCode, message, error]
      properties:
        statusCode:
          type: integer
          example: 404
        message:
          description: Single message, or an array of messages for class-validator errors
          oneOf:
            - type: string
            - type: array
              items:
                type: string
          example: File not found
        error:
          type: string
          example: Not Found

    # ── Auth ──
...
    FileResponse:
      description: >-
        Shape of the `File` Prisma model as returned by the API. `passwordHash`
        is stripped server-side and replaced by the `hasPassword` boolean —
        the raw hash is never sent to the client.
      type: object
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
          nullable: true
          description: null for anonymously-uploaded files
        originalName:
          type: string
          example: report-q4.pdf
        storageKey:
          type: string
          example: 3f6e.../a1b2c3d4-report-q4.pdf
        mimeType:
          type: string
          nullable: true
          example: application/pdf
        sizeBytes:
          type: integer
          format: int64
          description: >-
            Serialized as a JSON number here (BigInt → Number via the global
            patch in main.ts). Note: some other endpoints (list, download info)
            serialize the same field as a string instead — not yet harmonized.
          example: 2048576
        expiresAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        isDeleted:
          type: boolean
        hasPassword:
          type: boolean
    FileListItem:
      description: Shape returned by GET /files (deliberately excludes passwordHash).
      type: object
      properties:
        id:
          type: string
          format: uuid
        originalName:
          type: string
          example: report-q4.pdf
        storageKey:
          type: string
        mimeType:
          type: string
          nullable: true
        sizeBytes:
          type: string
          description: Serialized as a string here (`.toString()` in the service).
          example: "2048576"
        expiresAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        hasPassword:
          type: boolean
    FileStatsResponse:
      type: object
      properties:
        totalFiles:
          type: integer
        activeFiles:
          type: integer
        expiredFiles:
          type: integer
        totalSizeBytes:
          type: string
          example: "10485760"
    SetPasswordRequest:
      type: object
      required: [password]
      properties:
        password:
          type: string
          minLength: 4
          example: s3cr3t
    ManageTagsRequest:
      description: Replaces the full set of tags on the file (not additive).
      type: object
      required: [tags]
      properties:
        tags:
          type: array
          maxItems: 10
          items:
            type: string
            maxLength: 30
          example: ["design", "client-x"]
    TagsResponse:
      type: object
      properties:
        tags:
          type: array
          items:
            type: string
          example: ["design", "client-x"]
    DownloadHistoryEntry:
      type: object
      properties:
        id:
          type: string
          format: uuid
        downloadedAt:
          type: string
          format: date-time
        ipAddress:
          type: string
          nullable: true
        userAgent:
          type: string
          nullable: true
        tokenId:
          type: string
          format: uuid
          nullable: true

    # ── Download links (extension of US01) ──
...
paths:
...
  /files/upload:
    post:
      tags: [Files]
      summary: Upload a file (US01)
      description: |
        Upload a file with multipart/form-data. Does NOT create a download link —
        call POST /files/{id}/links afterwards to generate one.
      operationId: uploadFile
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
                  description: File to upload (max MAX_FILE_SIZE_BYTES, default 1 GB)
                expiryDays:
                  type: integer
                  minimum: 1
                  maximum: 30
                  default: 7
                  description: Expiration in days (defaults to FILE_EXPIRY_DAYS_DEFAULT)
                password:
                  type: string
                  description: Optional password protecting the file at upload time (US09)
      responses:
        "201":
          description: File uploaded successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FileResponse"
        "400":
          description: Validation error (no file, file too large, forbidden extension)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /files/anonymous:
    post:
      tags: [Files]
      summary: Upload a file anonymously (US07)
      description: |
        Upload a file without authentication. Not linked to any user account,
        cannot be listed or managed afterwards, fixed 24h expiration.
      operationId: uploadFileAnonymous
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
                  description: File to upload (max MAX_FILE_SIZE_BYTES, default 1 GB)
      responses:
        "201":
          description: File uploaded successfully (anonymous, expires in 24h)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FileResponse"
        "400":
          description: Validation error (no file, file too large, forbidden extension)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /files/stats:
    get:
      tags: [Files]
      summary: Aggregate file statistics for the authenticated user (US06)
      operationId: getFileStats
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Aggregate stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FileStatsResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /files:
    get:
      tags: [Files]
      summary: List user's files (US05)
      description: >-
        Returns all non-deleted files owned by the authenticated user, newest first.
        No pagination, sorting or tag-filtering query parameters are implemented yet.
      operationId: listFiles
      security:
        - BearerAuth: []
      responses:
        "200":
          description: List of user's files
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FileListItem"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /files/{fileId}:
    get:
      tags: [Files]
      summary: Get file metadata
      operationId: getFileMetadata
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/FileId"
      responses:
        "200":
          description: File metadata
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FileResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Not the file owner
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: File not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    delete:
      tags: [Files]
      summary: Delete a file (US06)
      description: |
        Soft-deletes the file, removes the object from MinIO storage, and
        invalidates all its download tokens. Only the file owner can delete.
      operationId: deleteFile
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/FileId"
      responses:
        "204":
          description: File deleted successfully
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Not the file owner
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: File not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /files/{fileId}/password:
    put:
      tags: [Files]
      summary: Set or replace a file's password (US09)
      operationId: setFilePassword
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/FileId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetPasswordRequest"
      responses:
        "200":
          description: Password set
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Password set
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Not the file owner
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: File not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    delete:
      tags: [Files]
      summary: Remove a file's password (US09)
      operationId: removeFilePassword
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/FileId"
      responses:
        "204":
          description: Password removed
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Not the file owner
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: File not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /files/{fileId}/tags:
    put:
      tags: [Files]
      summary: Replace a file's tags (US08)
      description: Replaces the full set of tags on the file (not additive).
      operationId: setFileTags
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/FileId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ManageTagsRequest"
      responses:
        "200":
          description: Tags updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TagsResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Not the file owner
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: File not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    get:
      tags: [Files]
      summary: Get a file's tags (US08)
      operationId: getFileTags
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/FileId"
      responses:
        "200":
          description: Tags for this file
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TagsResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Not the file owner
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: File not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /files/{fileId}/history:
    get:
      tags: [Files]
      summary: Get a file's per-download history log (extra feature, not one of the 10 official user stories)
      operationId: getFileHistory
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/FileId"
      responses:
        "200":
          description: Download history entries, newest first
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DownloadHistoryEntry"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Not the file owner
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: File not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  # ════════════════════════════════════════════
  # DOWNLOAD LINKS (extension of US01, authenticated)
  # ════════════════════════════════════════════
...